requirements.txt
amqp==2.5.1
anyjson==0.3.3
asn1crypto==1.0.1
attrs==19.2.0
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
Django==1.5.12
future==0.18.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
kombu==4.6.5
lxml==4.4.1
more-itertools==7.2.0
oauthlib==3.1.0
parsel==1.5.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-twitter==3.5
pytz==2019.3
queuelib==1.5.0
redis==3.3.10
requests==2.22.0
requests-oauthlib==1.2.0
Scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
Twisted==19.7.0
urllib3==1.25.6
vine==1.3.0
w3lib==1.21.0
zipp==0.6.0
zope.interface==4.6.0
